+++ src/asp/Tools/stereo_tri.cc
@@ -570,9 +570,28 @@ void disp_or_matches_work(std::string const& output_prefix,
     unalign_disparity(is_map_projected, disparity_maps[0], left_trans, right_trans,  
                       opt, unaligned_disp_file);
   }
-      
+  
+  // If the images are mapprojected and we know the original image names,
+  // use those for the match file. That because the matches are between
+  // the original images, not the map-projected ones.
+  std::string img_file_key = "INPUT_IMAGE_FILE"; 
+  std::string left_image = opt.in_file1, right_image = opt.in_file2;
+  {
+    std::string img_file;  
+    boost::shared_ptr<vw::DiskImageResource> rsrc(new vw::DiskImageResourceGDAL(opt.in_file1));
+    vw::cartography::read_header_string(*rsrc.get(), img_file_key, img_file);
+    if (!img_file.empty()) 
+      left_image = img_file;
+  }
+  {
+    std::string img_file;
+    boost::shared_ptr<vw::DiskImageResource> rsrc(new vw::DiskImageResourceGDAL(opt.in_file2));
+    vw::cartography::read_header_string(*rsrc.get(), img_file_key, img_file);
+    if (!img_file.empty())
+      right_image = img_file; 
+  }
   std::string match_file = ip::match_filename(output_prefix + "-disp",
-                                              opt.in_file1, opt.in_file2);
+                                              left_image, right_image);
       
   // Pull matches from disparity.
   if (stereo_settings().num_matches_from_disparity > 0 &&